Remove GenServers that provided a useless abstraction and level of indirection
If the goal is at some point to keep track of some state related to the
operations made within the pipeline, this can be done e.g. by broadcasting
special messages on an internal PubSub to which those servers would be
subscribed.

Concurrency (which we didn't really have anyway and is no worse without
those servers whose `call` were blocking) can be achieved via GenStage or
through simple tasks